requirements.txt
numpy
tqdm
pyyaml
numba>=0.53.1
--find-links https://data.pyg.org/whl/torch-1.10.0+cu113.html 
torch_scatter
Cython
scipy
dropblock
plyfile